Support for expanding directory paths.
This commit adds a new `expanddir` flag that allows to add files with matched extensions in specified directory paths. It also includes an option to add files within subdirectories recursively.

The `expand_directory()` function in the newly created `path` module is the key to achieve this feature, and the `execute_from_command_line()` function is changed to suport the new `expanddir` flag.